dashboard-app/backend/requirements.txt
flask==3.1.1
flask-cors==6.0.1
flask-sqlalchemy==3.1.1
pyodbc==5.2.0
python-dotenv==1.1.0